requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
pymongo>=3.12.0
opencv-python>=4.5.3
qrcode>=7.3
geopy>=2.2.0
numpy>=1.19.3
pillow>=8.3.2
python-dotenv>=0.19.0
requests>=2.31.0
python-jose>=3.3.0
